Refactored rest of corejs code
