+++ _dev/js/theme/components/password/usePasswordPolicy.js
@@ -1,4 +1,17 @@
 import { sprintf } from 'sprintf-js';
+import prestashop from 'prestashop';
+
+/**
+ * Verify password score.
+ * Estimate guesses needed to crack the password.
+ * @param {String} password
+ * @returns {Promise}
+ */
+window.prestashop.checkPasswordScore = async (password) => {
+  const zxcvbn = (await import('zxcvbn')).default;
+
+  return zxcvbn(password);
+};
 
 const { passwordPolicy: PasswordPolicyMap } = prestashop.selectors;
 
@@ -43,7 +56,6 @@ const watchPassword = async (
   feedbackContainer,
   hints,
 ) => {
-  const { prestashop } = window;
   const passwordValue = elementInput.value;
   const elementIcon = feedbackContainer.querySelector(PasswordPolicyMap.requirementScoreIcon);
   const result = await prestashop.checkPasswordScore(passwordValue);
