--- _dev/js/theme/corejs/index.js
@@ -1,11 +1,0 @@
-import $ from 'jquery';
-
-import './selectors';
-
-import { psShowHide } from './common';
-import initEmailFields from './email-idn';
-
-$(() => {
-  psShowHide();
-  initEmailFields('input[type="email"]');
-});
